Add modules and data loaders for book and users
