+++ data/users.json
@@ -1,0 +1,30 @@
+[
+  {
+    "id": 447,
+    "name": "abd",
+    "age": 43,
+    "parent_permission": true,
+    "classroom": "asd",
+    "rentals": [
+      {
+        "date": "234",
+        "book_title": "sad",
+        "book_author": "zxc"
+      }
+    ]
+  },
+  {
+    "id": 696,
+    "name": "wer",
+    "age": 0,
+    "parent_permission": true,
+    "specialization": "doc",
+    "rentals": [
+      {
+        "date": "4545",
+        "book_title": "yu",
+        "book_author": " nm"
+      }
+    ]
+  }
+]
